add isl_*_list_size as an alternative name for isl_*_list_n_*
The type of the element is implicit in the type of the list,
so it does not need to be specified explicitly in the name of the function.
Moreover, several other isl_*_list functions already omit this element type
in their names.

The *_size name matches that of isl_vec_size.
It somewhat conflicts with the name of isl_set_size,
but that function is not documented so it should not affect
the choice of the name of documented functions.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>