isl_stream_read_map: fix handling of empty optional formula
An empty optional formula should not affect the map being parsed,
but this was broken in isl-0.07-39-g62808ce5b4 (isl_stream_read_map:
use affine expressions during parsing, Mon Jul 25 09:36:00 2011 +0200).
Prior to this commit, the caller of read_disjuncts would intersect
the result with the constraints derived from the tuples.
Since this commit, these constraints are passed to read_disjuncts instead
and returned back together with the parsed disjuncts.
The case of an empty formula failed to be adjusted accordingly.
Simply return the original constraints instead of returning a universe map.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>