isl_scheduler.c: graph_find_node: handle NULL input
If graph_find_node is given a NULL space, then it would return
a NULL node, which would be indistinguishable from a missing node.
Some callers explicitly check for a NULL space before calling
graph_find_node, but some others do not.
Move the handling of NULL spaces into graph_find_node and
make it return an invalid node (pointing beyond the sequence of nodes)
in case of a missing node.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>