Adapt coscheduled_source to handle error conditions returned by level_before
When isl-0.14-306-g5e596b8c2f started to use the isl schedule tree to
detect the order of accesses, it introduced a level_before function that
may also return -1, but failed to introduce code for handling such error
conditions in the callers. This resulted in unspecified behavior and
crashes, in cases where level_before returned a failure, e.g. in case of
a compute out. When isl-0.18-753-g9963b91 introduced code to handle
coscheduled dependences it did not take into account the possible error
condition, similarly how all existing code was not aware of this error
condition. All code already released in isl-0.18 has been corrected in
isl-0.18-80-g1a47003. This commit takes care of the necessary changes
in coscheduled_source.

Signed-off-by: Tobias Grosser <tobias@grosser.es>
Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>