isl_scheduler.c: isl_sched_graph: keep track of original dependence graph
That is, instead of simply recording whether an isl_sched_graph
object represents an original dependence graph, record
the actual original dependence graph from which the graph has
been split off.  In the isl_sched_graph represents
an original dependence graph, then it points to itself.

This will be useful for handling node pointers that point
to the original dependence graph.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>