isl_scheduler.c: inter_dim_map: handle NULL input
It is not clear if any of these inputs can be NULL
in the current code base, but since the implementation depends
on these inputs not being NULL, it is best to check anyway.
In particular, if graph->lp is NULL, then the allocated isl_dim_map
will be too small for the subsequent operations.

Reported-by: Michael Kruse <MichaelKruse@meinersbur.de>
Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>