cpp: fix initialization of local variable holding pointer to lambda function
The name of the lambda function method argument was assumed to be "fn",
while the generation of the method header in
cpp_generator::print_method_header uses the name of the argument
in the C interface, which may not be "fn".
This problem does not currently have any visible effect because
all callback arguments of exported functions happen to be called "fn".
Use the correct name anyway.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>