Complete isl_aff to set interface
Add the function isl_aff_gt_set. There is no obvious
difference to all the existing isl_aff and isl_pw_aff functions that
perform the very same operation with a different comparison operator.

Add the function isl_aff_ne_set. The operation not-equal is special
in comparison to the other operations that
work on basic_sets, as its result may require two disjuncts to be
represented and is consequently not a basic_set any more. Hence, only
the function isl_aff_ne_set can be introduced, but not the function
isl_aff_ne_basic_set.

Explicitly document the behavior in presence of NaN elements.
In consistence with the recently documented behavior (isl-0.18-804-gdda9608)
of isl for isl_val, comparisons with NaN are always false. Hence,
the result set of the newly introduced functions does not contain elements
for which a comparison with NaN is involved.

Signed-off-by: Tobias Grosser <tobias@grosser.es>
Reviewed-by: Nicolas Vasilache <nicolas.vasilache@gmail.com>
Reviewed-by: Oleksandr Zinenko <oleksandr.zinenko@inria.fr>
Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>