isl_map_coalesce: extend handling of inequality adjacent to equality
The original support for an inequality constraint adjacent
to an equality constraint would look for a pair consisting
of this constraint and the corresponding equality constraint
that is adjacent to the inequality constraint.
However, if the second disjunct satisfies multiple
equality constraints, then this adjacent equality constraint
may be a linear combination of the explicit equality constraints
that is not itself explicitly available.
It will therefore not be detected.

Specifically check for such an inequality constraint even
if the corresponding equality constraint was not found.

Requested-by: Annanay Agarwal <cs14btech11001@iith.ac.in>
Tested-by: Michael Kruse <MichaelKruse@meinersbur.de>
Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>