add isl_union_map_every_map
This function checks whether the callback function returns true
on each map in the union map and stops checking as soon as
one of the calls returns false.

In principle, the same effect can be obtained by using
isl_union_map_foreach_map, keeping track of the results
separately and returning isl_stat_error as soon as processing
can stop.  Even though this mechanism is commonly used
inside isl, it does constitute a slight abuse of the exception
handling and should not be exposed to the user.
The new isl_union_map_every_map exposes the desired functionality
in a more obvious way.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>