isl_tab_basic_set_non_trivial_lexmin: postpone rollback until entering level
This is a minor clean-up.
In the original code, the rollback was performed while leaving
the level that backtracks to the previous level.  Now it is
performed on entering that previous level.
This makes the code slightly easier to understand because
the rollback is now performed at the level that created the snapshot.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>