isl_mat_rank: extract out hermite_first_zero_col
It would have been possible to use isl_mat_initial_non_zero_cols instead,
which is essentially identical, but hermite_first_zero_col will
be updated to exploit the column echelon form in the next commit.
It is also a bit more general (allowing the caller to specify
the first column to look at as well as the size of the columns),
which will allow it to be used several times in upcoming commits.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>