isl_union_map.c: isl_un_op_control: allow user data in filter callback
The next commit will use a filter callback that needs to pass in extra data.

The original filters (which do not require extra data) are handled
through an extra indirection.  An alternative would be to allow
for two types of filters, one with extra argument and one without,
but this would lead to near duplicate code in un_entry.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>