add isl_schedule_node_every_descendant
This function checks whether the callback function returns true
on each descendant of a node and stops checking as soon as
one of the calls returns false.

In principle, the same effect can be obtained by using
isl_schedule_node_foreach_descendant_top_down, keeping track
of the results separately and returning isl_bool_error as soon as
processing can stop.  Even though this mechanism is commonly used
inside isl and is in fact used inside the implementation
of isl_schedule_node_every_descendant, it does constitute
a slight abuse of the exception handling and should not be exposed
to the user.
The new isl_schedule_node_every_descendant exposes
the desired functionality in a more obvious way.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>