isl_schedule_constraints_compute_schedule: only sort statements when needed
At the very end of the scheduling algorithm, the statements are sorted
based on the remaining dependences.  In some cases, however,
all dependences have already been covered by the outer schedule nodes,
in which case there is no need to insert a sequence node
to order the statements.

Instead of sorting statements as soon as the schedule is complete,
only do so when the current band is empty.  This introduces
an extra round of checking for weakly connected components and
if each statement turns out to lie in a separate component,
then no further nodes need to be added.

This also helps in grouping statements together that actually
do still depend on each other as different such groups
will get separated apart by a set node.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>