drop deprecated isl_int interface
The isl_int interface was deprecated in isl-0.12-11-g6272dbd9df
(deprecate isl_int, Fri May 3 18:31:59 2013 +0200).
Most of the formerly exposed functions are still used internally.
The functions that are not used in the current code base and
that are unlikely to be used in the near future are removed
together with the declarations.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>