isl_output.c: print_disjunct: pass valid local variables to print_constraints
The original implementation would pass a NULL object in case
no explicit representations of local variables should be printed.
Pass in a valid object with all local variables marked unknown instead.
This ensures that print_affine will always get passed a valid object,
which is required for the next commit.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>