isl_scheduler.c: non_neg_lexmin: fix double free on error path
The error path frees both "sol" and "prev", but since
isl-0.18-679-g6e75a0dd15 (scheduler: solve Feautrier fallback as ILP
to avoid large coefficients, Wed Apr 12 10:49:44 2017 +0200),
this error path can be reached with both pointing to the same object
(without an increased reference count).
Do not overload "prev", used for storing the previous solution, with also
indicating that another solution should be computed, but use a separate
variable for this indication.  This allows the scope of "prev"
to be reduced such that it is always either NULL or pointing to
a different object.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>