isl_test.c: test_un_val: do not consider internal failure as incorrect result
In particular, there are two ways for a test to fail,
either the computation of the result failed or the wrong result was computed.
In the second case, an error message is printed, but there is
no need to so in the first case as some error message should have
already been printed at the point where the computation failed.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>