add exported isl_union_map_intersect_range_wrapped_domain_union_set
This function allows the wrapped domain to be intersected directly,
without having to first modify the union map to expose the domain.
Similarly to isl_union_map_intersect_domain_wrapped_domain_union_set
in the previous commit, this function has a type suffix so that
a space variant could be added later.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>