isl_stream_eat: print error message on EOF
isl_stream_eat is called in a situation where a specific token
is expected and it already prints an error message when
a different token is found.  The callers therefore do
not print any (extra) error message.
However, in case of EOF, no error message was printed,
resulting in an error condition without any indication
of what went wrong.
Print this error message.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>