isl_*_aff_read_from_str: do not read domains as rational sets
When support for parsing rational sets was added in
isl-0.10-176-gce84f3f800 (isl_stream_read_map: add partial support
for reading rational sets, Sun Jul 15 09:52:53 2012 +0200),
the reading of domains of affine expressions was changed to
read those domains as rational sets (without motivation).
While the expressions themselves may be rational, there is
no need to assume that the domain is rational.
Furthermore, when printing affine functions, a rational domain
is not marked in any way, resulting in an affine function
that looks identical to one with an integral domain but
that is different, as, for example, in the new test case.

Read the domains as integral sets to avoid this confusion.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>