isl_tab: invalidate undo stack if pushing record fails
The user of the isl_tab may still try to rollback the tableau
even after some operation failed.
If the last record did not get added, then this may result
in spurious errors of the type "not undoing last redundant row".
Invalidate the undo stack to ensure no actual rollback will be performed.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>