export more isl_schedule_node functionality
Functions that are not exported are those that require support
for extra types or features such as isl_id and isl_space.
The function isl_schedule_node_band_sink is not exported because
the generator assumes that an object returned by a subclass
method is of the subclass type, which may not be the case
for isl_schedule_node_band_sink.

Reviewed-by: Tobias Grosser <tobias.grosser@inf.ethz.ch>
Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>