add isl_union_pw_aff_param_on_domain
This is a convenience wrapper around isl_union_pw_aff_aff_on_domain
that does not require the user to manually set up an isl_aff object.

The internally used isl_aff_param_on_domain_space_id function is not
exposed for now, but it could be exposed later if there is a need.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>