add isl_space_add_param_id
This convenience function allows parameters to be added
to a space without having to resort to first adding an unnamed
parameter and then changing its identifier.
Unnamed parameters and changing the names/identifiers of parameters
should in general be discouraged.
This function will also be used in the next commit.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>