drop C++ bindings
The bindings currently being generated are those for use
in the specific case where exceptions cannot be used.
Feedback on a proposal for also generating C++ bindings
with exceptions revealed that there is no consensus
on what these two sets of bindings should look like,
or indeed on whether they could somehow be combined.
Drop the current C++ bindings such that this issue
does not block the 0.19 release.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>