add exported isl_set_unbind_params{,_insert_domain}
During the construction of a polyhedral model, it is not uncommon
for a statement instance set to be constructed from a description
in terms of the parameters at the point of the statement (including
the outer loop iterators) and a list of outer loop iterators
that will become the set dimensions of the instance set.
Prior to this commit, the only way of transforming
a description in terms of the parameters into one where
some of these parameters are reinterpreted as the set dimensions,
is to call isl_set_move_dims.  However, the use of this function
is discouraged.
isl_set_unbind_params provides a more structured and
hopefully also more natural way of achieving this effect.

isl_set_unbind_params_insert_domain performs a similar
operation, but creates a map from a set rather than a set
from a parameter domain.
This function can be used, for example, to create access relations
in the same setting as described above.

An earlier design took a tuple identifier and a list of parameters
as two separate arguments instead of the isl_multi_id.
This turned out to be less convenient as it requires the user
to keep track of two pieces of information.
The current design is also more general as it also works
with structured spaces.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>