add isl_multi_*_get_list
These functions extract the constituent base expressions
from a multiple expression.  They can already be extracted
individually using isl_multi_*_get_at, but sometimes
it is more convenient to get a hold of them as a whole.

A naming of the form isl_multi_*_get_*_list would
be more similar to isl_multi_*_get_* and isl_multi_*_from_*_list,
but there is a trend towards not including the return type
in the function name if this return type is implicit
in the type of the object on which the function is called,
as in, for example, isl_multi_*_get_at.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>