add exported isl_multi_id
This type of objects keeps track of a space along with
identifiers for the set dimensions.  This will be useful
for transforming parameters into set dimensions with
the identifiers specifying the parameters that need
to be transformed into the set dimensions at the corresponding
positions and vice versa.

While a space itself can also have names for set dimensions,
these are transient and only meant for pretty printing.
They are not guaranteed to survive any operation
(on the space itself or indeed on any other object somehow
related to the space).

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>