add exported isl_multi_*_size
It was already possible to obtain the number of base expressions
in a multiple expression, but only through a rather awkward interface.
In particular, it is also possible to obtained this number
by calling isl_multi_*_dim.  However, the dimension type that
needs to be specified depends on whether the multi expression
lives in a set or a map space.  Furthermore, for such a basic
property of a multi expression, the user should not have
to bother with any dimension type.

It is not entirely obvious what would be the best name
for this function.  Since a multi-expression is
a (possibly structured) tuple, valid names
would appear to be "length" or "size".
This commit picks "size".

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>