add isl_map_get_range_stride_info
This function is similar to isl_set_get_stride_info, except that
it produces an offset in terms of the input dimensions instead
of one in terms of the other dimensions.
The main reason that isl_map_get_range_stride_info cannot
return an offset in terms of all dimensions is that it
is currently impossible to represent an affine function
with a map domain (unless it is wrapped).
It also turns out to be most useful in practice.

Reviewed-by: Oleksandr Zinenko <oleksandr.zinenko@inria.fr>
Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>