Add isl_*_drop_unused_params
Add functions to drop parameters that are not referenced by the isl object.

When using named parameters and (align_params), isl objects tend to
accumulate many parameter dimensions over time. To improve the
performance on later operations (due to working on lower dimensional
objects) and because the printed sets are more readable, it is useful to
be able to drop unused parameter dimensions.

These functions require all parameters to be named as otherwise
it would be difficult to tell what the meaning is
of the remaining parameters.

Signed-off-by: Tobias Grosser <tobias@grosser.es>
Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>