add convenience functions for adding a tuple to a parameter or a set space
When a set space is created from a parameter space,
the user would have to call two to three functions before,
one to create an unnamed zero-dimensional set space,
one to set the dimension and (optionally)
one to set the name of the space.
Provide convenience functions that only require a single call.
Similarly for creating a map space from a domain space.

Reviewed-by: Manjunath Kudlur <manjunath@cerebras.net>
Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>