isl_multi_union_pw_aff_range_product: fix handling of missing explicit domains
In particular, if isl_multi_union_pw_aff_range_product was given
two zero-dimensional arguments, one with explicit domain and
one without, then if it is the first one that has the explicit domain,
it would be preserved, while in the other case it would not.
The reason it that while isl_multi_*_intersect_explicit_domain
can rely on isl_multi_*_domain_intersect to handle the first case,
it must handle the second case itself.
Add this handling to isl_multi_*_intersect_explicit_domain.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>