make isl_multi_*_drop_dims declaration optional
Conceptually, this declaration belongs to ISL_DECLARE_MULTI_DIMS,
but it is also defined for isl_multi_union_pw_aff, while the other
declarations in ISL_DECLARE_MULTI_DIMS are not.
Similarly for isl_multi_*_dim, which is needed in order
to be able to call isl_multi_*_drop_dims.

While these two functions will get defined for isl_multi_id
(at least for now), there is no need to expose them in the public API.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>