isl_reordering_free: return isl_reordering * instead of void *
While the currently returned void pointer is automatically converted
to "isl_reordering *" where needed, it is both clearer and more in
line with other isl functions to return a pointer to the type of the
object that is freed. This allows analysis tools to catch bugs where the
return type is implicitly converted to a different type.

Signed-off-by: Tobias Grosser <tobias@grosser.es>
Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>