isl_cell_foreach_simplex: use isl_stat instead of int for local variable
The only value assigned to the variable 'r' is of type isl_stat and the
return type of the function which returns 'r' is also of type isl_stat.
This change adjusts the type of the local variable to isl_stat to avoid
two implicit conversions between int and isl_stat. This is an oversight
from 657fb59f (isl_vertices_foreach_{disjoint_cell,simplex}: return
isl_stat, Tue Oct 25 11:50:07 2016 +0200).  Fixing it improves
readability and helps static analysis tools.

Signed-off-by: Tobias Grosser <tobias@grosser.es>
Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>