isl_set_get_stride: remove local variables from local space of offset
The offset is derived from a constraint with integer divisions,
with the integer divisions removed from the affine expression.
However, those integer divisions were left in the local space
of the affine expression, resulting in the offset not being
obviously equal to the same affine expression without those
extra variables in the local space.
Remove them.

Reported-by: Oleksandr Zinenko <oleksandr.zinenko@inria.fr>
Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>