add isl_multi_union_pw_aff_{min,max}_multi_val
The motivation for the naming and the interface of these functions
is the same as for isl_union_pw_aff_{min,max}_val
introduced in an earlier commit.

Unlike those functions, which could have been easily
implemented in terms of the original internals of
isl_union_set_min_multi_union_pw_aff, it would be
a bit more involved to do the same for
isl_multi_union_pw_aff_min_multi_val because
each of the multiple expressions may have a different domain and
the union of these domains would have had to be collected
to be used as a dummy argument to isl_multi_union_pw_aff_min_multi_val.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>