checked C++ bindings: abort on unchecked boolean error when converting to bool
The purpose of the checked bindings is to ensure that all
error conditions within an entire region of code are handled.
That is, the program should act on a return value that indicates
an error.  In the case of a method returning an isl::boolean,
this means that the calling code should check whether
the return value satisfies is_error().  Failing to do so would
mean that the code could abort when run in the wild,
completely defeating the purpose of the checking bindings.
In particular, if it is OK for this region to abort, then
the standard bindings would do.

Ideally, the check for errors should be enforced by the compiler,
but this looks non-trivial and is left for future work.

The next best thing is to detect missing error checks at run-time,
whenever the code is executed.  Only detecting the missing error
check when an actual error occurs, as the current bindings do,
makes it *much* harder to detect missing error checks since
the occurrence of an error during a particular call depends on
the input data and the history of all previous calls to isl
in the entire program.  This significantly increasing the risk
that the code will abort when run in the wild, which is
exactly the kind of situation that should be avoided.

Improve on the current situation by detecting missing error checks
whenever the code is run and not only when an actual error occurs.
In particular, complain whenever an isl::checked:boolean
is converted to bool without a prior check for errors.
If is_error() is called, then there clearly has been such a check.
A call to is_false() or isl_true() is also taken a check for errors
since these function returning true means that no error has occurred.
The assumption here is that the user would tend to write is_false()
instead of !is_true().

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>