add isl_pw_multi_aff_get_range_tuple_id convenience function
The interface of isl_pw_multi_aff_get_tuple_id is not well designed
because it only works for some values of the "type" argument.
Moreover, the value in theory depends on whether
the function lives in a set (isl_dim_set) or a map space (isl_dim_out).
In practice, isl_dim_set is equal to isl_dim_out, but this knowledge
should only be used inside isl.
Provide an isl_pw_multi_aff_get_range_tuple_id alternative
to avoid these issues.
Note that only multi-functions can have a range tuple identifier,
so this function is not added for other isl_pw_* types.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>