add exported isl_space_get_{domain,range}_tuple_id convenience functions
The interface of isl_space_get_tuple_id is not well designed
because it only works for some values of the "type" argument.
Moreover, the value of this argument depends on whether
the first argument is a set or a map space.
Provide isl_space_get_{domain,range}_tuple_id alternatives
to avoid these issues.
In particular, these functions should only be called on map spaces.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>