no longer use whole-component scheduler by default
The whole-component scheduler is the original scheduler introduced in
isl-0.06-43-g1192654a2 (add isl_union_set_compute_schedule,
Sat Mar 19 14:47:47 2011 +0100).  In isl-0.16.1-11-ge3e81206b
(isl_schedule_constraints_compute_schedule: allow incremental combination
of SCCs, Tue Apr 14 09:59:19 2015 +0200) an incremental scheduler
was added, along with an option to select the original, whole-component
scheduler.  This option was turned on by default so as not to break
compatibility.  The incremental scheduler was, however, made the default
in PPCG and later also in Tensor Comprehensions.
Since the incremental scheduler is more flexible, it is probably best
to use it by default in new projects.  It's therefore best
to stop using the whole-component scheduler by default in isl.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>