C++ bindings: throw exceptions of correct exception subclass
The purpose of the subclasses is for the user to be able
to catch different classes of isl exceptions separately,
but this ability somehow got lost while preparing the commit
for upstreaming.  Make it work (again) and make sure it keeps working
by adjusting the test to catch specific subclasses.

Reported-by: Philip Pfaffe <philip.pfaffe@gmail.com>
Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>