introduce isl_val_pow2 as an alternative name for isl_val_2exp
Functions that have names starting with integer literals rather than
letters are invalid both for the python and for the C++ bindings.
This commit adds isl_val_pow2 as an alternative name for isl_val_2exp
to make it possible to expose this function in the python and C++ bindings.

While it would have been possible to rename the function during bindings
generation similar to the union -> unite rename already performed, for
consistency between the different bindings and the C interface itself it
is preferable to use a better function name. (This was not possible in
the 'union/unite' case as the term 'union' is indeed preferable and
works well for both C and python).

Signed-off-by: Tobias Grosser <tobias@grosser.es>
Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>