python bindings: also raise an exception on negative isl_stat
The only currently exported functions returning isl_stat
are the "foreach" functions.  If such a function returns
with an error condition, then it is usually due to an error
that occurred inside the callback function and this case
is already handled.  However, the error could also occur
outside the callback function and then an exception needs
to be raised as well.  On the other hand, there is no need
to return anything from methods generated from such functions.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>