isl_map_coalesce: do not leave input basic maps in inconsistent state
Commit isl-0.11.1-19-g9ce5772769 (isl_map_coalesce: avoid dropping
constraints redundant wrt implicit equalities,
Sat Mar 16 14:40:56 2013 +0100) introduced code to ensure
that equality constraints discovered in the tableau are
explicitly available in the basic map, without performing
Gaussian elimination on those equality constraints during
the isl_map_coalesce call, postponing this step to right before
the result is returned.  However, it would only end up performing
this elimination on the basic maps that remain in the result and
not on those that get dropped and that may be shared with other maps.
Also perform Gaussian elimination on those basic maps.

An alternative would be to only make the equality constraints explicit
in a copy of the input basic maps, but discovering equality constraints
takes some effort, so it seems best to not simply throw this information
away in case the basic maps are shared.

Reported-by: Andrei Lascu <alas.20073@gmail.com>
Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>